potcar

+++ bin/potcar.sh
@@ -1,24 +1,33 @@
-#! /bin/bash
+#!/bin/bash
 
 # Create POSTCAR
 # To use it: postcar.sh Fe B O H
+# OR postcar.sh Without parameters, it will be automatically extracted from POSCAR
 
 # Define local potpaw_GGA pseduopotential repository:
-
 #repo="/home/lineng/vasp/user/libing/VASP/potpaw_PBE1"
 
 # 查找脚本的目录
 repo=${LB_HOME}"/pbe"
 #echo $repo
 
+if [ ! -f POSCAR ]; then
+	echo "### WARING: POSCAR NOT FOUND."
+	exit
+fi 
+
+str=`sed -n '6,1p' POSCAR | tr '\r' ' ' | tr '\n' ' '`
+
+arr=($str)
+
 # Check if older version of POTCAR is present
 if [ -f POTCAR ] ; then
 	mv -f POTCAR old-POTCAR
-	echo "### WARING: old POSTCAR file found and renamed to 'old-POTCAR'."
+	# echo "### WARING: old POSTCAR file found and renamed to 'old-POTCAR'."
 fi
 
 # Main loop - concatenate the appropriatePOTCARs (or archives)
-for i in $*
+for i in ${arr[*]};
 do
  	if test -f $repo/$i/POTCAR ; then
 		cat $repo/$i/POTCAR>> POTCAR
@@ -27,7 +36,7 @@ do
 	elif test -f $repo/$i/POTCAR.gz ; then
 		gunzip -c $repo/$i/POTCAR.gz >> POTCAR
 	else
-		echo "### WARING: No suitable POTCAR for element '$i' found!! Skipped thiselement."
+		echo "### WARING: No suitable POTCAR for element $i !!! Skipped this element."
  	fi
 done
